+++ src/main/resources/db/migration/V3__create_table_lotes.sql
@@ -1,6 +1,6 @@
 create table lotes(
     id BIGSERIAL primary key,
-    quantidade DECIMAL,
+    quantidade INTEGER,
     data_validade DATE,
     data_entrada DATE,
     produto_id BIGINT references produto (id)
